requirements.txt
numpy==1.11.1
Pillow==3.3.1
-e git+git@github.com:BradyKieffer/SimplexNoise.git@52efe256d82392815f05cbc5c59d32a04ed653b6#egg=SimplexNoise
